Tests/LibWeb/Text/expected/URL/invalid-urls.txt

new URL('file://xn--/p', undefined)
error creating URL: 'TypeError: Invalid URL'
new URL('http://0xffffffff1', undefined)
error creating URL: 'TypeError: Invalid URL'
